+++ fabric-chaincode-shim/src/main/java/org/hyperledger/fabric/shim/impl/ChaincodeSupportClient.java
@@ -22,7 +22,7 @@ import java.util.logging.Logger;
 
 public class ChaincodeSupportClient {
     private static final int DEFAULT_TIMEOUT = 5;
-    private static Logger logger = Logger.getLogger(ChaincodeSupportClient.class.getName());
+    private static final Logger LOGGER = Logger.getLogger(ChaincodeSupportClient.class.getName());
     private static Logger perflogger = Logger.getLogger(Logging.PERFLOGGER);
     private final ManagedChannel channel;
     private final ChaincodeSupportStub stub;
@@ -35,7 +35,11 @@ public class ChaincodeSupportClient {
         this.stub = ChaincodeSupportGrpc.newStub(channel);
     }
 
-    private void shutdown(final InnvocationTaskManager itm) {
+    /**
+     *
+     * @param itm
+     */
+    public void shutdown(final InnvocationTaskManager itm) {
 
         // first shutdown the thread pool
         itm.shutdown();
@@ -56,76 +60,12 @@ public class ChaincodeSupportClient {
     }
 
     /**
-     * @param itm
-     */
-    public void start(final InnvocationTaskManager itm) {
-
-        // This is a critical method - it is the one time that a
-        // protobuf service is invoked. The single 'register' call
-        // is made, and two streams are created.
-        //
-        // It is confusing how these streams are then used to send messages
-        // to and from the peer.
-        //
-        // the response stream is the message flow FROM the peer
-        // the 'request observer' is the message flow TO the peer
-        //
-        // Messages coming from the peer will be requests to invoke
-        // chaincode, or will be the responses to stub APIs, such as getState
-        // Message to the peer will be the getState APIs, and the results of
-        // transaction invocations
-
-        // The InnvocationTaskManager's way of being told there is a new
-        // message, until this is received and processed there is now
-        // knowing if this is a new transaction function or the answer to say getState
-        final Consumer<ChaincodeMessage> consumer = itm::onChaincodeMessage;
-
-        logger.info("making the grpc call");
-        // for any error - shut everything down
-        // as this is long lived (well forever) then any completion means something
-        // has stopped in the peer or the network comms, so also shutdown
-        final StreamObserver<ChaincodeMessage> requestObserver = this.stub.register(
-
-                new StreamObserver<ChaincodeMessage>() {
-                    @Override
-                    public void onNext(final ChaincodeMessage chaincodeMessage) {
-                        // message off to the ITM...
-                        consumer.accept(chaincodeMessage);
-                    }
-
-                    @Override
-                    public void onError(final Throwable t) {
-                        logger.severe(() -> "An error occured on the chaincode stream. Shutting down the chaincode stream." + Logging.formatError(t));
-
-                        ChaincodeSupportClient.this.shutdown(itm);
-                    }
-
-                    @Override
-                    public void onCompleted() {
-                        logger.severe("Chaincode stream is complete. Shutting down the chaincode stream.");
-                        ChaincodeSupportClient.this.shutdown(itm);
-                    }
-                }
-
-        );
-
-        try {
-            start(itm, requestObserver);
-        } catch (IOException e) {
-            // befor external chaincode start method not handle NullPointerException
-            e.printStackTrace();
-            throw new RuntimeException(e);
-        }
-    }
-
-    /**
+     *
      * @param itm
      * @param requestObserver
-     * @return
      * @throws IOException verify parameters error
      */
-    public StreamObserver<ChaincodeShim.ChaincodeMessage> start(final InnvocationTaskManager itm,
-                                                                final StreamObserver<ChaincodeMessage> requestObserver) throws IOException {
+    public void start(final InnvocationTaskManager itm, final StreamObserver<ChaincodeMessage> requestObserver) throws IOException {
         if (requestObserver == null) {
             throw new IOException("StreamObserver 'requestObserver' for chat with peer can't be null");
         }
@@ -163,25 +103,13 @@ public class ChaincodeSupportClient {
         // NOTE the register() - very important - as this triggers the ITM to send the
         // first message to the peer; otherwise the both sides will sit there waiting
         itm.setResponseConsumer(consumer).register();
+    }
 
-        return new StreamObserver<ChaincodeMessage>() {
-            @Override
-            public void onNext(final ChaincodeMessage chaincodeMessage) {
-                itm.onChaincodeMessage(chaincodeMessage);
-            }
-
-            @Override
-            public void onError(final Throwable t) {
-                logger.severe(() -> "An error occured on the chaincode stream. Shutting down the chaincode stream." + Logging.formatError(t));
-
-                ChaincodeSupportClient.this.shutdown(itm);
-            }
-
-            @Override
-            public void onCompleted() {
-                logger.severe("Chaincode stream is complete. Shutting down the chaincode stream.");
-                ChaincodeSupportClient.this.shutdown(itm);
-            }
-        };
+    /**
+     * ChaincodeSupportStub.
+     * @return
+     */
+    public ChaincodeSupportStub getStub() {
+        return stub;
     }
 }
